refactor(configuration): own the TraceDecay config model
