+++ crates/tracedecay-cli/src/commands/index.rs
@@ -21,7 +21,7 @@ async fn is_fresh_install() -> bool {
 
 /// When invoked with no subcommand, offer to create the index if none exists.
 pub(crate) async fn handle_no_command() -> tracedecay_domain::errors::Result<()> {
-    let project_path = tracedecay::config::resolve_path(None);
+    let project_path = tracedecay_configuration::resolve_path(None);
     if TraceDecay::has_initialized_store(&project_path).await {
         // Already initialized — show help via clap
         let _ = <crate::cli::Cli as clap::CommandFactory>::command().print_help();
@@ -78,7 +78,7 @@ pub(crate) async fn handle_init(
     fresh: bool,
     assume_yes: bool,
 ) -> tracedecay_domain::errors::Result<()> {
-    let project_path = tracedecay::config::resolve_path(path);
+    let project_path = tracedecay_configuration::resolve_path(path);
     let profile_root = tracedecay_runtime_core::storage::default_profile_root()?;
     if let Some(message) =
         tracedecay_global_db::ephemeral_root_rejection(&project_path, &profile_root)
@@ -497,9 +497,10 @@ pub(crate) async fn handle_sync(
             message: "brokered sync does not yet support --skip-folders/--include-folders; update tracedecay.toml first".to_string(),
         });
     }
-    let resolved =
-        super::scope::resolve_project_scope(tracedecay::config::resolve_path_with_discovery(path))
-            .await?;
+    let resolved = super::scope::resolve_project_scope(
+        tracedecay_configuration::resolve_path_with_discovery(path),
+    )
+    .await?;
     let handshake = tracedecay::daemon::handshake_for_current_client(
         Some(resolved.project_path.clone()),
         None,
