+++ crates/tracedecay-configuration/Cargo.toml
@@ -8,7 +8,9 @@ description = "Transport-neutral configuration control plane and runtime pin sur
 repository = "https://github.com/ScriptedAlchemy/tracedecay"
 
 [dependencies]
+glob = "0.3"
 hotpath.workspace = true
+serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 sha2 = "0.11"
 thiserror = "2"
@@ -17,6 +19,8 @@ tracedecay-contracts = { path = "../tracedecay-contracts", version = "0.1.0" }
 tracedecay-automation = { path = "../tracedecay-automation", version = "0.1.0" }
 tracedecay-domain = { path = "../tracedecay-domain", version = "0.1.0" }
 tracedecay-global-db = { path = "../tracedecay-global-db", version = "0.1.0" }
+tracedecay-lcm = { path = "../tracedecay-lcm", version = "0.1.0" }
+tracedecay-maintenance = { path = "../tracedecay-maintenance", version = "0.1.0" }
 tracedecay-policy = { path = "../tracedecay-policy", version = "0.1.0" }
 tracedecay-runtime-core = { path = "../tracedecay-runtime-core", version = "0.1.0" }
 tracedecay-semantic-contracts.workspace = true
