+++ crates/tracedecay-configuration/src/lib.rs
@@ -7,11 +7,16 @@
 pub mod config;
 pub mod configuration;
 
+pub use config::model::{
+    CONFIG_FILENAME, MIN_AUTO_TRACK_PR_POLL_SECS, RetentionConfig, SYNC_RETENTION_SETTING_KEY,
+    SyncConfig, TelemetryConfig, TraceDecayConfig, brand_env, get_config_path, is_excluded,
+    is_excluded_dir, is_generated_path_segment, is_in_gitignore, is_included, load_config,
+    load_config_from_path, resolve_path, resolve_path_with_discovery, save_config_to_path,
+};
 pub use config::{
     OpenedRuntimeConfiguration, PinnedRuntimeConfiguration, PinnedRuntimeConfigurationCachePort,
-    RuntimeConfigurationTarget, SyncConfig, TelemetryConfig, TraceDecayConfig,
-    cached_pinned_runtime_configuration, install_pinned_runtime_configuration_cache,
-    publish_pinned_runtime_configuration,
+    RuntimeConfigurationTarget, cached_pinned_runtime_configuration,
+    install_pinned_runtime_configuration_cache, publish_pinned_runtime_configuration,
 };
 pub use configuration::{
     AuthorizedActor, CONFIGURATION_AUDIT_PAGE_LIMIT, ComponentConfigurationState,
