+++ crates/tracedecay/src/daemon/doctor_kernel.rs
@@ -341,7 +341,7 @@ async fn collect_over_budget_store_findings(
         tracedecay_contracts::storage::StoreKeyV1,
         GuardedStoreTelemetryPort,
     )],
-    retention: &crate::config::RetentionConfig,
+    retention: &tracedecay_configuration::RetentionConfig,
 ) -> CollectedStoreTelemetryV1 {
     use std::collections::BTreeMap;
     use tracedecay_contracts::storage::{
@@ -789,7 +789,7 @@ pub(in crate::daemon) fn production_doctor_report_reader(
     profile_root: PathBuf,
     host_home: Option<PathBuf>,
     remote_operational: RemoteOperationalReadProviderV1,
-    retention: crate::config::RetentionConfig,
+    retention: tracedecay_configuration::RetentionConfig,
     schedulers: tracedecay_code_index_runtime::code_index_scheduler::CodeIndexSchedulerRegistryV1,
     diagnostic_broker: Arc<tokio::sync::Mutex<tracedecay_lsp::analyzer::broker::DiagnosticBroker>>,
     feedback_runtimes: DaemonFeedbackRuntimeRegistrar,
