+++ crates/tracedecay/src/mcp/server.rs
@@ -489,7 +489,7 @@ pub struct McpServer {
     /// The `[sync]` config resolved once at construction from the project
     /// root (plus `TRACEDECAY_SYNC_*` env overrides). Cached so the read
     /// hot path never re-reads the config file per `tools/call`.
-    sync_config: crate::config::SyncConfig,
+    sync_config: tracedecay_configuration::SyncConfig,
     /// Savings-ledger recorder tasks spawned so far / finished so far, plus
     /// a notifier pinged on every completion. Production never awaits these
     /// (ledger writes stay fire-and-forget); tests await
@@ -1289,7 +1289,7 @@ impl McpServer {
         }
     }
 
-    pub(crate) fn watcher_sync_config(&self) -> &crate::config::SyncConfig {
+    pub(crate) fn watcher_sync_config(&self) -> &tracedecay_configuration::SyncConfig {
         &self.sync_config
     }
