+++ crates/tracedecay/src/tracedecay/queries/meta.rs
@@ -1,7 +1,7 @@
 use std::path::Path;
 
-use crate::config::TraceDecayConfig;
 use crate::tracedecay::TraceDecay;
+use tracedecay_configuration::TraceDecayConfig;
 use tracedecay_domain::errors::{Result, TraceDecayError};
 
 fn parse_counter(key: &'static str, value: Option<String>) -> Result<u64> {
